+++ tests/RFC4180FieldTest.php
@@ -3,10 +3,7 @@
 /**
  * League.Csv (https://csv.thephpleague.com).
  *
- * @author  Ignace Nyamagana Butera <nyamsprod@gmail.com>
- * @license https://github.com/thephpleague/csv/blob/master/LICENSE (MIT License)
- * @version 9.2.0
- * @link    https://github.com/thephpleague/csv
+ * (c) Ignace Nyamagana Butera <nyamsprod@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -20,9 +17,9 @@ use League\Csv\RFC4180Field;
 use League\Csv\Writer;
 use PHPUnit\Framework\TestCase;
 use TypeError;
+use function stream_get_filters;
 use const STREAM_FILTER_ALL;
 use const STREAM_FILTER_READ;
-use function stream_get_filters;
 
 /**
  * @group filter